REF: share NumericArray._arith_method with BooleanArray
