+++ pandas/core/arrays/numeric.py
@@ -23,19 +23,15 @@ from pandas.core.dtypes.common import (
     is_bool_dtype,
     is_float_dtype,
     is_integer_dtype,
-    is_list_like,
     is_object_dtype,
     is_string_dtype,
     pandas_dtype,
 )
 
-from pandas.core import ops
-from pandas.core.arrays.base import ExtensionArray
 from pandas.core.arrays.masked import (
     BaseMaskedArray,
     BaseMaskedDtype,
 )
-from pandas.core.construction import ensure_wrapped_if_datetimelike
 
 if TYPE_CHECKING:
     import pyarrow
@@ -213,65 +209,6 @@ class NumericArray(BaseMaskedArray):
 
         scalars = to_numeric(strings, errors="raise")
         return cls._from_sequence(scalars, dtype=dtype, copy=copy)
-
-    def _arith_method(self, other, op):
-        op_name = op.__name__
-        omask = None
-
-        if isinstance(other, BaseMaskedArray):
-            other, omask = other._data, other._mask
-
-        elif is_list_like(other):
-            if not isinstance(other, ExtensionArray):
-                other = np.asarray(other)
-            if other.ndim > 1:
-                raise NotImplementedError("can only perform ops with 1-d structures")
-
-        # We wrap the non-masked arithmetic logic used for numpy dtypes
-        #  in Series/Index arithmetic ops.
-        other = ops.maybe_prepare_scalar_for_op(other, (len(self),))
-        pd_op = ops.get_array_op(op)
-        other = ensure_wrapped_if_datetimelike(other)
-
-        mask = self._propagate_mask(omask, other)
-
-        if other is libmissing.NA:
-            result = np.ones_like(self._data)
-            if "truediv" in op_name and self.dtype.kind != "f":
-                # The actual data here doesn't matter since the mask
-                #  will be all-True, but since this is division, we want
-                #  to end up with floating dtype.
-                result = result.astype(np.float64)
-        else:
-            # Make sure we do this before the "pow" mask checks
-            #  to get an expected exception message on shape mismatch.
-            if self.dtype.kind in ["i", "u"] and op_name in ["floordiv", "mod"]:
-                # ATM we don't match the behavior of non-masked types with
-                #  respect to floordiv-by-zero
-                pd_op = op
-
-            with np.errstate(all="ignore"):
-                result = pd_op(self._data, other)
-
-        if op_name == "pow":
-            # 1 ** x is 1.
-            mask = np.where((self._data == 1) & ~self._mask, False, mask)
-            # x ** 0 is 1.
-            if omask is not None:
-                mask = np.where((other == 0) & ~omask, False, mask)
-            elif other is not libmissing.NA:
-                mask = np.where(other == 0, False, mask)
-
-        elif op_name == "rpow":
-            # 1 ** x is 1.
-            if omask is not None:
-                mask = np.where((other == 1) & ~omask, False, mask)
-            elif other is not libmissing.NA:
-                mask = np.where(other == 1, False, mask)
-            # x ** 0 is 1.
-            mask = np.where((self._data == 0) & ~self._mask, False, mask)
-
-        return self._maybe_mask_result(result, mask, other, op_name)
 
     _HANDLED_TYPES = (np.ndarray, numbers.Number)
