+++ pandas/tests/arrays/masked/test_arithmetic.py
@@ -153,22 +153,29 @@ def test_error_len_mismatch(data, all_arithmetic_operators):
 
     other = [scalar] * (len(data) - 1)
 
+    err = ValueError
     msg = "|".join(
         [
             r"operands could not be broadcast together with shapes \(3,\) \(4,\)",
             r"operands could not be broadcast together with shapes \(4,\) \(3,\)",
         ]
     )
-
-    if data.dtype.kind == "b":
-        msg = "Lengths must match"
+    if data.dtype.kind == "b" and all_arithmetic_operators.strip("_") in [
+        "sub",
+        "rsub",
+    ]:
+        err = TypeError
+        msg = (
+            r"numpy boolean subtract, the `\-` operator, is not supported, use "
+            r"the bitwise_xor, the `\^` operator, or the logical_xor function instead"
+        )
 
     for other in [other, np.array(other)]:
-        with pytest.raises(ValueError, match=msg):
+        with pytest.raises(err, match=msg):
             op(data, other)
 
         s = pd.Series(data)
-        with pytest.raises(ValueError, match=msg):
+        with pytest.raises(err, match=msg):
             op(s, other)
